+++ src/components/Sidebar/FooterItem/index.ts
@@ -1,1 +1,1 @@
-export * from './FooterItem';
+export * from './FooterItemList';
