+++ src/components/Sidebar/NavItem/index.ts
@@ -1,1 +1,1 @@
-export * from './NavItem';
+export * from './NavItemList';
